requirements.txt
matplotlib==3.8.4
numpy==2.2.2
pyhocon==0.3.61
pypose==0.6.8
pyproj==3.7.0
PyYAML==6.0.2
PyYAML==6.0.2
scipy==1.15.1
simplekml==1.3.6
torch==2.5.1
tqdm==4.66.5
wandb==0.19.4